Begin refactoring and cleaning code base for new ui
