+++ happiness-frontend/src/App.js
@@ -1,31 +1,27 @@
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import Statistics from "./pages/Statistics";
 import Profile from "./pages/Profile";
-import Header from "./components/Header";
 import Sidebar from "./components/Sidebar";
-import Welcome from "./pages/Welcome";
+import Welcome from "./pages/authentication/Welcome";
 import Container from "react-bootstrap/Container";
 import Settings from "./pages/Settings";
 import History from "./pages/History";
-import UserGroups from "./pages/UserGroups";
-import Group from "./pages/Group";
+import Group from "./pages/groups";
 import "./App.css";
 import ApiProvider from "./contexts/ApiProvider";
 import UserProvider from "./contexts/UserProvider";
 import PrivateRoute from "./components/PrivateRoute";
 import PublicRoute from "./components/PublicRoute";
-import SubmitHappiness from "./pages/SubmitHappiness";
 import RequestResetPassword from "./pages/authentication/RequestResetPassword";
 import ResetPassword from "./pages/authentication/ResetPassword";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
-import Entries from "./pages/Entries";
 import NewUserGroups from "./pages/NewUserGroups";
 import { StyledEngineProvider } from "@mui/material";
+import Entries from "./pages/happiness/CommentHeader";
 
 export default function App() {
-  const USE_NEW_UI = process.env.REACT_APP_USE_NEW_UI;
   const bgStyle =
-    "mx-auto mt-5 mb-3 ms-4 me-4 " + (!USE_NEW_UI ? " max-w-7xl" : "");
+    "mx-auto mt-5 mb-3 ms-4 me-4  max-w-7xl";
   const theme = createTheme({
     typography: {
       fontFamily: ["Inter", "cursive"].join(","),
@@ -36,7 +32,7 @@ export default function App() {
     <Routes>
       <Route
         path="/home"
-        element={USE_NEW_UI ? <Entries /> : <SubmitHappiness />}
+        element={<Entries /> }
       />
       <Route
         path="/statistics"
@@ -58,7 +54,7 @@ export default function App() {
         path="/groups"
         element={
           <div className={bgStyle}>
-            {USE_NEW_UI ? <NewUserGroups /> : <UserGroups />}
+            {<NewUserGroups />}
           </div>
         }
       />
@@ -126,14 +122,7 @@ export default function App() {
                     path="*"
                     element={
                       <PrivateRoute>
-                        {USE_NEW_UI ? (
                           <Sidebar element={privateRoutes} />
-                        ) : (
-                          <>
-                            <Header />
-                            {privateRoutes}
-                          </>
-                        )}
                       </PrivateRoute>
                     }
                   />
