+++ happiness-frontend/src/pages/groups/index.js
@@ -1,99 +1,15 @@
-import { useParams } from "react-router-dom";
-import { Fragment, useEffect, useState } from "react";
-import { Tab } from "@headlessui/react";
-import GroupManage from "../components/groups/GroupManage";
-import { useApi } from "../contexts/ApiProvider";
-import { useQuery } from "react-query";
-import TableView from "../components/groups/TableView";
-import GroupStats from "../components/groups/GroupStats";
-import { GetRangeHappiness } from "../components/happinessHistory/GetHappinessData";
-import { Spinner } from "react-bootstrap";
-
-function TabButton({ text }) {
-  return (
-    <Tab as={Fragment}>
-      {({ selected }) => (
-        <button
-          className={
-            "inline-block px-2 py-2.5 m-1.5 w-[140px] rounded-lg " +
-            (selected
-              ? "text-cultured-50 bg-raisin-600"
-              : "hover:text-gray-900 hover:bg-gray-100 dark:hover:bg-gray-800 dark:hover:text-white")
-          }
-        >
-          {text}
-        </button>
-      )}
-    </Tab>
-  );
-}
-
-// Title for tab with buttons to change the currently viewed time range
-// size is the text size of the main tab
-export function TimeButtonTitle({
-  text,
-  radioValue,
-  setStart,
-  setEnd,
-  size = "sm:text-3xl text-2xl",
-}) {
-  return (
-    <div className="flex items-center justify-center">
-      {/* View earlier time period button */}
-      <button
-        type="button"
-        className="px-3 me-3 w-[50px] h-[40px] rounded-lg text-cultured-50 bg-raisin-600 text-xl"
-        onClick={() => {
-          setStart((start) => {
-            if (radioValue === 1) {
-              start.setDate(start.getDate() - 7);
-            } else {
-              start.setMonth(start.getMonth() - 1);
-            }
-            return new Date(start);
-          });
-          setEnd((end) => {
-            if (radioValue === 1) {
-              end.setDate(end.getDate() - 7);
-            } else {
-              end.setMonth(end.getMonth() - 1);
-            }
-            return new Date(end);
-          });
-        }}
-      >
-        &lt;
-      </button>
-      {/* Title text */}
-      <p className={"font-medium m-3 text-raisin-600 " + size}>{text}</p>
-      {/* View later time period button */}
-      <button
-        type="button"
-        className="px-3 ms-3 w-[50px] h-[40px] rounded-lg text-cultured-50 bg-raisin-600 text-xl"
-        onClick={() => {
-          setStart((start) => {
-            if (radioValue === 1) {
-              start.setDate(start.getDate() + 7);
-            } else {
-              start.setMonth(start.getMonth() + 1);
-            }
-            return new Date(start);
-          });
-          setEnd((end) => {
-            if (radioValue === 1) {
-              end.setDate(end.getDate() + 7);
-            } else {
-              end.setMonth(end.getMonth() + 1);
-            }
-            return new Date(end);
-          });
-        }}
-      >
-        &gt;
-      </button>
-    </div>
-  );
-}
+import {useParams} from "react-router-dom";
+import {Fragment, useEffect, useState} from "react";
+import {Tab} from "@headlessui/react";
+import GroupManage from "../../components/groups/GroupManage";
+import {useApi} from "../../contexts/ApiProvider";
+import {useQuery} from "react-query";
+import TableView from "../../components/groups/TableView";
+import GroupStats from "../../components/groups/GroupStats";
+import {GetRangeHappiness} from "../../components/happinessHistory/GetHappinessData";
+import {Spinner} from "react-bootstrap";
+import {TabButton} from "./TabButton";
+import {TimeButtonTitle} from "./TimeButtonTitle";
 
 export default function Group() {
   const { groupID } = useParams();
