+++ happiness-frontend/src/util/Formatting.js
@@ -1,0 +1,10 @@
+export function formatDate(date) {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, "0");
+    const day = String(date.getDate()).padStart(2, "0");
+    return `${year}-${month}-${day}`;
+}
+
+export function formatHappinessNum(happiness) {
+    return (Math.round(happiness * 2) / 2).toFixed(1);
+}
